refactor: Extract scale domain inference and normalization into Scale methods
Major refactoring to improve code organization and maintainability:

- Moved scale domain inference logic from render.rs into Scale::infer_domain_from_data()
  - Handles DataFields evaluation and conversion to band scales for string data
  - Evaluates domain expressions using DataFusion

- Created Scale::normalize_domain() to apply zero/nice/padding transformations
  - Ensures axes and data use the same normalized scale domains
  - Fixes scale synchronization issues between axes and marks

- Added Scale::create_configured_scale() to encapsulate ConfiguredScale creation
  - Consolidates domain/range/options conversion logic
  - Applies scale defaults based on usage context

- Simplified PlotRenderer::render() by ~150 lines
  - Replaced inline domain inference with method calls
  - Improved readability and maintainability

- Added centralized scale defaults system (scales/defaults.rs)
  - Context-aware defaults (e.g., zero=true for y-axis linear scales)
  - Allows user overrides while providing sensible defaults

- Fixed scale domain inference for expression-based encodings
  - Creates derived columns for complex expressions like col("x").sub(lit(2.0))
  - Properly computes min/max from transformed values

- Updated visual test baselines for minor rendering differences
  - Fixed scatter plot alignment issues with nicified domains
  - All tests passing with improved accuracy

This refactoring makes the scale system more modular and easier to extend,
while fixing several bugs related to scale normalization and expression handling.

+++ avenger-chart/src/coords.rs
@@ -56,8 +56,8 @@ impl CoordinateSystem for Cartesian {
 
     fn default_range(&self, channel: &str, width: f64, height: f64) -> Option<ScaleRange> {
         match channel {
-            "x" => Some(ScaleRange::new_interval(lit(0.0), lit(width))),
-            "y" => Some(ScaleRange::new_interval(lit(height), lit(0.0))), // Inverted for screen coords
+            "x" => Some(ScaleRange::new_interval(lit(0.0f32), lit(width as f32))),
+            "y" => Some(ScaleRange::new_interval(lit(height as f32), lit(0.0f32))), // Inverted for screen coords
             _ => None,
         }
     }
@@ -115,12 +115,15 @@ impl CoordinateSystem for Polar {
     fn default_range(&self, channel: &str, width: f64, height: f64) -> Option<ScaleRange> {
         match channel {
             "theta" => Some(ScaleRange::new_interval(
-                lit(0.0),
-                lit(2.0 * std::f64::consts::PI),
+                lit(0.0f32),
+                lit((2.0 * std::f64::consts::PI) as f32),
             )),
             "r" => {
                 let max_radius = f64::min(width, height) / 2.0;
-                Some(ScaleRange::new_interval(lit(0.0), lit(max_radius)))
+                Some(ScaleRange::new_interval(
+                    lit(0.0f32),
+                    lit(max_radius as f32),
+                ))
             }
             _ => None,
         }
